Extract subset of a sipnet clim file based on start and end time

Extract subset of a sipnet clim file based on start and end time

Usage

split_sipnet_met(start.time, stop.time, met, overwrite = FALSE, outpath = NULL)

Arguments

start.time

start date and time at which to start extraction

stop.time

stop date and time at which to stop extraction

met

path to sipnet clim file to be split

overwrite

if TRUE, overwrite existing target file (Default FALSE)

outpath

if specified, write output to a new directory. Default NULL writes back to the directory being read

Value

path to split up climate file

Author

Mike Dietze, Ann Raiho, Alexey Shiklomanov