Writes a configuration files for SIPNET model

Writes a configuration files for your model

Usage

write.config.SIPNET(
  defaults,
  trait.values,
  settings,
  run.id,
  inputs = NULL,
  IC = NULL,
  restart = NULL,
  spinup = NULL
)

Arguments

defaults

pft

trait.values

vector of samples for a given trait

settings

PEcAn settings object

run.id

run ID

inputs

list of model inputs

IC

initial condition

restart

In case this is a continuation of an old simulation. restart needs to be a list with name tags of runid, inputs, new.params (parameters), new.state (initial condition), ensemble.id (ensemble id), start.time and stop.time.See Details.

spinup

currently unused, included for compatibility with other models

Author

Michael Dietze